refactor: :recycle: moves summary cards to own folder
