+++ frontend/src/components/Agreements/AgreementDetailsCards/AgreementTotalCard/index.js
@@ -1,0 +1,1 @@
+export { default } from "./AgreementTotalCard";
