+++ frontend/src/components/Agreements/AgreementDetailsCards/BLIsByFYSummaryCard/BLIsByFYSummaryCard.styles.module.scss
@@ -8,10 +8,6 @@
     border-width: 1px;
     border-bottom-left-radius: 4px;
     border-top-left-radius: 4px;
-
-    //&:hover {
-    //    transform: scaleY(1.4);
-    //}
 }
 
 .rightBar {
@@ -19,13 +15,4 @@
     border-width: 1px;
     border-bottom-right-radius: 4px;
     border-top-right-radius: 4px;
-
-    //&:hover {
-    //    transform: scaleY(1.4);
-    //}
-}
-
-.rightBarFull {
-    //border-bottom-left-radius: 4px;
-    //border-top-left-radius: 4px;
 }
