+++ frontend/src/components/Agreements/AgreementDetailsCards/BLIsByFYSummaryCard/index.js
@@ -1,0 +1,1 @@
+export { default } from "./BLIsByFYSummaryCard";
